+++ src/test/java/org/mmarini/railways2/model/TrainTest.java
@@ -41,6 +41,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mmarini.railways2.model.RailwayConstants.COACH_LENGTH;
 
 class TrainTest {
+    public static final int GAME_DURATION = 300;
     StationStatus status;
     StationMap stationMap;
 
@@ -54,7 +55,7 @@ class TrainTest {
                 .addNode("b", new Point2D.Double(200, 0), "ab")
                 .addTrack("ab", "a", "b")
                 .build();
-        status = new StationStatus.Builder(stationMap, 1, null)
+        status = new StationStatus.Builder(stationMap, 1, GAME_DURATION, null)
                 .addRoute(Entry::create, "a")
                 .addRoute(Exit::create, "b")
                 .build();