+++ CHANGELOG.md
@@ -20,7 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue #50: Add mouse event on map panel
 - Issue #53: Compute distance from edges
 - Issue #56: Add status change by user interaction
-  
+- Issue #58: Add entry/exit train counters
+
 ### Changed 
 
 - Issue #3: Stop at platform for loading